+++ packages/sdoc-editor/src/extension/constants/index.js
@@ -94,6 +94,15 @@ export const HEADERS = [
   HEADER6,
 ];
 
+export const HEADER_TAG = {
+  header1: 'H1',
+  header2: 'H2',
+  header3: 'H3',
+  header4: 'H4',
+  header5: 'H5',
+  header6: 'H6',
+};
+
 export const HEADER_TITLE_MAP = {
   [TITLE]: 'Title',
   [SUBTITLE]: 'Subtitle',